[Llama3-70b] Separate vllm generator class and add prompt length validation in input processor
